feature: player - 이름판단 추가
