+++ src/main/java/step2/domain/player/Player.java
@@ -2,16 +2,29 @@ package step2.domain.player;
 
 import java.util.Objects;
 
+import step2.util.ErrorTarget;
+import step2.util.Validator;
+
 public class Player {
 
-	private final Name name;
+	private static final int MAX_NAME_LENGTH = 5;
+
+	private final String name;
 
 	public Player(String input) {
-		this.name = new Name(input);
+		Validator.notBlank(input, ErrorTarget.NAME_INPUT);
+		Validator.max(MAX_NAME_LENGTH, input.length(),
+			String.format("이름이 최대 길이는 %d 글자 입니다. 입력 : %s", MAX_NAME_LENGTH, input));
+
+		this.name = input;
+	}
+
+	public boolean hasName(String playerName) {
+		return this.name.equals(playerName);
 	}
 
 	public String getName() {
-		return name.getValue();
+		return name;
 	}
 
 	@Override
